Folder restructure and removed lib.rs
