+++ src/logic/rule_parser.rs
@@ -1,13 +1,7 @@
-use crate::*;
-use permutation_iter::PermutationIter;
+use Token::*;
 
 use anyhow::{anyhow, Context, Result};
-use std::borrow::Borrow;
-use std::collections::{HashMap, HashSet};
-use std::fmt;
 use std::iter::Peekable;
-use std::rc::Rc;
-use Token::*;
 
 #[derive(Debug, Copy, Clone)]
 pub enum Direction {
@@ -128,141 +122,8 @@ impl<'a> RuleParser {
     }
 }
 
-// TruthTable struct holds the truth table data of an input rule.
-// It can be constructed using a PermutationIter. Since the permutations generated by
-// PermuationIter is always guaranteed to follow the same pattern, the order of the results
-// implicitly holds the propositional data for each entry in the TruthTable. Example:
-// `0 => 0` implies index 0b00, results[0]
-// `0 => 1` implies index 0b01, results[1]
-// `1 => 0` implies index 0b10, results[2]
-// `1 => 1` implies index 0b11, results[3]
-#[derive(Default, Eq, PartialEq, Hash, Clone)]
-pub struct TruthTable {
-    pub variables: Vec<char>,
-    pub results: Vec<bool>,
-}
-
-impl TruthTable {
-    pub fn new() -> Self {
-        TruthTable {
-            variables: Vec::new(),
-            results: Vec::new(),
-        }
-    }
-}
-
-impl TryFrom<PermutationIter> for TruthTable {
-    type Error = anyhow::Error;
-
-    fn try_from(mut permutation_iter: PermutationIter) -> Result<Self, Self::Error> {
-        let mut table = Self::new();
-        let mut parser = RuleParser::new();
-        for permutation in permutation_iter.by_ref() {
-            table.results.push(
-                parser
-                    .evaluate(&permutation)
-                    .context(format!("Failed to evaluate permutation {}", permutation))?,
-            );
-        }
-        table.variables.append(&mut permutation_iter.variables);
-        Ok(table)
-    }
-}
-
-impl fmt::Debug for TruthTable {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let len = self.variables.len();
-        for v in &self.variables {
-            write!(f, "| {} ", v)?;
-        }
-        writeln!(f, "| = |")?;
-        writeln!(f, "{}|", "|---".repeat(len + 1))?;
-        for (i, result) in self.results.iter().enumerate() {
-            for b in 0..len {
-                write!(
-                    f,
-                    "| {} ",
-                    if i & (1 << (len - 1 - b)) == 0 { 0 } else { 1 }
-                )?
-            }
-            writeln!(f, "| {} |", if *result { 1 } else { 0 })?;
-        }
-        Ok(())
-    }
-}
-
-// Structure that holds key pairs of identifier and all related truth tables.
-#[derive(Default)]
-pub struct RuleMap {
-    map: HashMap<char, HashSet<Rc<TruthTable>>>,
-}
-
-impl RuleMap {
-    // Inserts a new rule in the rulemap Ad-Hoc
-    pub fn insert<T>(&mut self, rule: T) -> Result<()>
-    where
-        T: Borrow<str>,
-    {
-        let ptr = Rc::new(
-            TruthTable::try_from(PermutationIter::new(rule.borrow())).context(format!(
-                "Failed to create truth table from: '{}'",
-                rule.borrow()
-            ))?,
-        );
-        for v in ptr.variables.iter() {
-            let tables = self
-                .map
-                .entry(*v)
-                .or_insert_with(|| HashSet::from([Rc::clone(&ptr)]));
-            tables.insert(Rc::clone(&ptr));
-        }
-        Ok(())
-    }
-
-    pub fn insert_vec<T>(&mut self, rules: Vec<T>) -> Result<()>
-    where
-        T: Borrow<str>,
-    {
-        for rule in rules.iter() {
-            self.insert(rule.borrow())?
-        }
-        Ok(())
-    }
-}
-
-impl<T> TryFrom<Vec<T>> for RuleMap
-where
-    T: Borrow<str>,
-{
-    type Error = anyhow::Error;
-
-    fn try_from(rules: Vec<T>) -> Result<Self> {
-        let mut map = RuleMap::default();
-        map.insert_vec(rules)?;
-        Ok(map)
-    }
-}
-
-impl fmt::Debug for RuleMap {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let mut map = self.map.iter().peekable();
-        while let Some((k, v)) = map.next() {
-            writeln!(f, "{}", k)?;
-            let mut table = v.iter().peekable();
-            while let Some(t) = table.next() {
-                if map.peek().is_none() && table.peek().is_none() {
-                    write!(f, "{:?}", t)?;
-                } else {
-                    writeln!(f, "{:?}", t)?;
-                }
-            }
-        }
-        Ok(())
-    }
-}
-
 #[cfg(test)]
-mod tests_rule_parser {
+mod tests {
     use super::*;
 
     use anyhow::Result;
@@ -417,94 +278,6 @@ mod tests_rule_parser {
         assert_eq!(
             result.unwrap_err().to_string(),
             "Missing closing parenthesis"
-        );
-    }
-}
-
-#[cfg(test)]
-mod tests_truth_table {
-    use super::*;
-
-    use anyhow::Result;
-    use pretty_assertions::assert_eq;
-
-    #[test]
-    fn simple() -> Result<()> {
-        let result = TruthTable::try_from(PermutationIter::new("A => Z"))?;
-        assert_eq!(result.variables, vec!['A', 'Z']);
-        assert_eq!(result.results, vec![true, true, false, true]);
-        Ok(())
-    }
-
-    #[test]
-    fn error_invalid_rule() {
-        let result = TruthTable::try_from(PermutationIter::new("A = Z"));
-        assert!(result.is_err());
-        assert_eq!(
-            result.unwrap_err().to_string(),
-            "Failed to evaluate permutation 0 = 0"
-        );
-    }
-}
-
-#[cfg(test)]
-mod tests_rule_map {
-    use super::*;
-
-    use anyhow::Result;
-    use pretty_assertions::assert_eq;
-
-    #[test]
-    fn empty() -> Result<()> {
-        let result = RuleMap::try_from(Vec::<String>::new())?;
-        assert!(result.map.is_empty());
-        Ok(())
-    }
-
-    #[test]
-    fn from() -> Result<()> {
-        let result = RuleMap::try_from(vec!["A => B", "B => C"])?;
-        assert_eq!(result.map.len(), 3);
-        assert_eq!(result.map.get(&'A').map_or(0, |v| v.len()), 1);
-        assert_eq!(result.map.get(&'B').map_or(0, |v| v.len()), 2);
-        assert_eq!(result.map.get(&'C').map_or(0, |v| v.len()), 1);
-        Ok(())
-    }
-
-    #[test]
-    fn insert() -> Result<()> {
-        let mut result = RuleMap::try_from(Vec::<String>::new())?;
-        result.insert("A => B")?;
-        assert_eq!(result.map.len(), 2);
-        assert_eq!(result.map.get(&'A').map_or(0, |v| v.len()), 1);
-        assert_eq!(result.map.get(&'B').map_or(0, |v| v.len()), 1);
-
-        result.insert("B => C")?;
-        assert_eq!(result.map.len(), 3);
-        assert_eq!(result.map.get(&'A').map_or(0, |v| v.len()), 1);
-        assert_eq!(result.map.get(&'B').map_or(0, |v| v.len()), 2);
-        assert_eq!(result.map.get(&'C').map_or(0, |v| v.len()), 1);
-        Ok(())
-    }
-
-    #[test]
-    fn insert_vec() -> Result<()> {
-        let mut result = RuleMap::try_from(Vec::<String>::new())?;
-        result.insert_vec(vec!["A => B", "B => C"])?;
-        assert_eq!(result.map.len(), 3);
-        assert_eq!(result.map.get(&'A').map_or(0, |v| v.len()), 1);
-        assert_eq!(result.map.get(&'B').map_or(0, |v| v.len()), 2);
-        assert_eq!(result.map.get(&'C').map_or(0, |v| v.len()), 1);
-        Ok(())
-    }
-
-    #[test]
-    fn error_invalid_rule() {
-        let result = RuleMap::try_from(vec!["A =>"]);
-        assert!(result.is_err());
-        assert_eq!(
-            result.unwrap_err().to_string(),
-            "Failed to create truth table from: 'A =>'"
         );
     }
 }
