+++ src/usage.rs
@@ -1,0 +1,8 @@
+//This is a stand alone file so that integration_test.rs can use this with a hardcoded path
+
+use indoc::indoc;
+
+pub const USAGE: &str = indoc! {"
+TODO: add usage
+
+"};
