+++ src/utils/mod.rs
@@ -1,0 +1,7 @@
+mod read_file;
+
+#[cfg(test)]
+#[path = "../../tests/test_utils/mod.rs"]
+pub mod test_utils;
+
+pub use read_file::*;
