+++ src/utils/read_file.rs
@@ -1,12 +1,4 @@
-pub mod input;
-pub mod parser;
-pub mod permutation_iter;
-pub mod sanitize;
-
 use anyhow::{anyhow, Context, Result};
-use core::borrow::Borrow;
-use core::char;
-use indoc::indoc;
 use std::{
     any::type_name,
     fs::File,
@@ -14,15 +6,6 @@ use std::{
     path::Path,
     str::FromStr,
 };
-
-pub const USAGE: &str = indoc! {"
-TODO: add usage
-
-"};
-
-pub fn is_identifier(c: impl Borrow<char>) -> bool {
-    ('A'..='Z').contains(c.borrow())
-}
 
 pub fn read_file<T: FromStr>(file: &impl AsRef<Path>) -> Result<Vec<T>> {
     let file = File::open(file).context("Failed to open file")?;
@@ -40,26 +23,9 @@ pub fn read_file<T: FromStr>(file: &impl AsRef<Path>) -> Result<Vec<T>> {
 }
 
 #[cfg(test)]
-#[path = "../tests/test_utils/mod.rs"]
-pub mod test_utils;
-
-#[cfg(test)]
-mod tests_is_identifier {
+mod tests {
     use super::*;
-
-    use pretty_assertions::assert_eq;
-
-    #[test]
-    fn all() {
-        for c in '\0'..=char::MAX {
-            assert_eq!(is_identifier(&c), ('A'..='Z').contains(&c));
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests_read_file {
-    use super::*;
+    use crate::utils::test_utils;
 
     use anyhow::Result;
     use pretty_assertions::assert_eq;
